+++ src/test/java/com/sprint/mission/discodeit/DiscodeitApplicationTests.java
@@ -2,12 +2,14 @@ package com.sprint.mission.discodeit;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
+@ActiveProfiles("test")
 class DiscodeitApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+  @Test
+  void contextLoads() {
+  }
 
 }
