+++ src/test/java/com/sprint/mission/discodeit/controller/integration/MessageControllerIntegrationTest.java
@@ -1,10 +1,10 @@
-package com.sprint.mission.discodeit.controller;
-
+package com.sprint.mission.discodeit.controller.integration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sprint.mission.discodeit.service.UserService;
+import com.jayway.jsonpath.JsonPath;
+import com.sprint.mission.discodeit.controller.MessageController;
+import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,21 +17,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
 
-// todo - webEnvironment 옵션
-// todo - api 통합 테스트
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Transactional
 @ActiveProfiles("test")
-public class UserControllerIntegrationTest {
-
-  @Autowired
-  private UserService userService;
-
-  @Autowired
-  private UserController userController;
-
-  @Autowired
-  private ObjectMapper objectMapper;
+class MessageControllerIntegrationTest {
 
   @Autowired
   private ApplicationContext context;
@@ -43,19 +32,22 @@ public class UserControllerIntegrationTest {
   private TestRestTemplate restTemplate;
 
   @Test
-  @DisplayName("유저 get api 통합 테스트")
-  void getUser_integrationTest() {
-
+  @DisplayName("메시지 get api 통합 테스트")
+  void getMessage_integrationTest() {
     // given
-    String url = "http://localhost:" + port + "/api/users";
+    UUID messageId = UUID.randomUUID();
+
+    String url = "http://localhost:" + port + "/api/messages?channelId=" + messageId;
 
     // when
     ResponseEntity<String> response = restTemplate.getForEntity(url, String.class);
 
     // then
+    String content = JsonPath.read(response.getBody(), "$.content").toString();
+
     assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-    assertThat(response.getBody()).isEqualTo("[]");
-    assertThat(context.getBean("userController")).isInstanceOf(UserController.class);
-    assertThat(context.getBean(UserController.class)).isNotNull();
+    assertThat(content).isEqualTo("[]");
+    assertThat(context.getBean("messageController")).isInstanceOf(MessageController.class);
+    assertThat(context.getBean(MessageController.class)).isNotNull();
   }
 }
